Move all components
