+++ components/date-picker/style/TimePicker.less
@@ -1,5 +1,3 @@
-@import "../../mixins/input";
-
 .@{calendar-prefix-cls}-time {
   // Change display order in DOM
   .@{calendar-prefix-cls}-input-wrap {
