+++ components/date-picker/style/index.less
@@ -1,5 +1,6 @@
 @import "../../../style/themes/default";
 @import "../../../style/mixins/index";
+@import "../../input/style/mixin";
 
 @calendar-prefix-cls: ant-calendar;
