+++ components/input/style/mixin.less
@@ -1,3 +1,6 @@
+@import "../../../style/mixins/index";
+
+@import "../../../style/mixins/index";
 // size mixins for input
 .input-lg() {
   padding: @input-padding-vertical-lg @input-padding-horizontal;
