+++ components/modal/style/index.js
@@ -1,2 +1,5 @@
 import '../../../style/core/index.less';
 import './index.less';
+
+// style dependencies
+import '../../button/style';
