+++ components/popconfirm/style/index.js
@@ -1,0 +1,2 @@
+import '../../../style/core/index.less';
+import '../../popover/style';
