+++ components/queue-anim/style.js
@@ -1,0 +1,1 @@
+// placeholder, no style
