+++ components/radio/style/index.less
@@ -1,0 +1,223 @@
+@import "../../../style/themes/default";
+@import "../../../style/mixins/index";
+
+@radio-prefix-cls: ant-radio;
+@radio-group-prefix-cls: ~"@{radio-prefix-cls}-group";
+@radio-inner-prefix-cls: ~"@{radio-prefix-cls}-inner";
+@radio-duration: .2s;
+
+.@{radio-group-prefix-cls} {
+  display: inline-block;
+  label {
+    margin-right: 16px;
+  }
+}
+
+// 一般状态
+.@{radio-prefix-cls} {
+  white-space: nowrap;
+  outline: none;
+  display: inline-block;
+  position: relative;
+  line-height: 1;
+  vertical-align: middle;
+  cursor: pointer;
+  &:hover {
+    .@{radio-inner-prefix-cls} {
+      border-color: #bcbcbc;
+    }
+  }
+  &-inner {
+    &:after {
+      position: absolute;
+      width: 6px;
+      height: 6px;
+      left: 3px;
+      top: 3px;
+      border-radius: @border-radius-base;
+      display: table;
+      border-top: 0;
+      border-left: 0;
+      content: ' ';
+      background-color: @primary-color;
+      opacity: 0;
+      transform: scale(0);
+      transition: all @radio-duration @ease-in-out-circ;
+    }
+
+    position: relative;
+    top: 0;
+    left: 0;
+    display: inline-block;
+    width: 14px;
+    height: 14px;
+    border-width: 1px;
+    border-style: solid;
+    border-radius: 14px;
+    border-color: @border-color-base;
+    background-color: #fff;
+    transition: all @radio-duration @ease-in-out-circ;
+  }
+
+  &-input {
+    position: absolute;
+    left: 0;
+    z-index: 1;
+    cursor: pointer;
+    opacity: 0;
+    top: 0;
+    bottom: 0;
+    right: 0;
+  }
+}
+
+// 选中状态
+.@{radio-prefix-cls}-checked {
+  .@{radio-inner-prefix-cls} {
+    border-color: @border-color-base;
+    &:after {
+      transform: scale(1);
+      opacity: 1;
+      transition: all @radio-duration @ease-in-out-circ;
+    }
+  }
+}
+
+.@{radio-prefix-cls}-disabled {
+  &:hover {
+    .@{radio-inner-prefix-cls} {
+      border-color: @border-color-base;
+    }
+  }
+  .@{radio-inner-prefix-cls} {
+    border-color: @border-color-base;
+    background-color: #f3f3f3;
+    &:after {
+      background-color: #cccccc;
+    }
+  }
+
+  .@{radio-inner-prefix-cls}-input {
+    cursor: default;
+  }
+}
+
+span.@{radio-prefix-cls} + * {
+  margin-left: 8px;
+}
+
+.@{radio-prefix-cls} {
+  label&-button {
+    background: #fff;
+    padding: 0 16px;
+    margin: 0;
+    height: 28px;
+    line-height: 26px;
+    color: #666;
+    display: inline-block;
+    transition: all 0.3s ease;
+    cursor: pointer;
+    border: 1px solid @border-color-base;
+    border-left: 0;
+
+    a {
+      color: #666;
+    }
+
+    > span.@{radio-prefix-cls}-button {
+      margin-left: 0;
+      display: block;
+      width: 0;
+      height: 0;
+    }
+
+    .@{radio-group-prefix-cls}-large & {
+      height: 32px;
+      line-height: 30px;
+    }
+
+    .@{radio-group-prefix-cls}-small & {
+      height: 22px;
+      line-height: 20px;
+      padding: 0 12px;
+      &:first-child {
+        border-radius: @border-radius-sm 0 0 @border-radius-sm;
+      }
+      &:last-child {
+        border-radius: 0 @border-radius-sm @border-radius-sm 0;
+      }
+    }
+
+    &:first-child {
+      border-radius: @border-radius-base 0 0 @border-radius-base;
+      border-left: 1px solid @border-color-base;
+    }
+
+    &:last-child {
+      border-radius: 0 @border-radius-base @border-radius-base 0;
+    }
+
+    &:first-child:last-child {
+      border-radius: @border-radius-base;
+    }
+
+    &:hover {
+      color: @primary-color;
+      position: relative;
+    }
+
+    .@{radio-prefix-cls}-inner,
+    input {
+      .opacity(0);
+      width: 0;
+      height: 0;
+    }
+
+    &-checked {
+      background: #fff;
+      border-color: @primary-color;
+      color: @primary-color;
+      box-shadow: -1px 0 0 0 @primary-color;
+
+      &:first-child {
+        border-color: @primary-color;
+        box-shadow: none!important;
+      }
+
+      &:hover {
+        border-color: tint(@primary-color, 20%);
+        box-shadow: -1px 0 0 0 tint(@primary-color, 20%);
+        color: tint(@primary-color, 20%);
+      }
+
+      &:active {
+        border-color: shade(@primary-color, 5%);
+        box-shadow: -1px 0 0 0 shade(@primary-color, 5%);
+        color: shade(@primary-color, 5%);
+      }
+    }
+
+    &-disabled {
+      border-color: @border-color-base;
+      background-color: #f7f7f7;
+      cursor: not-allowed;
+      color: #ccc;
+
+      &:first-child,
+      &:hover {
+        border-color: @border-color-base;
+        background-color: #f7f7f7;
+        color: #ccc;
+      }
+      &:first-child {
+        border-left-color: @border-color-base;
+      }
+    }
+
+    &-disabled.@{radio-prefix-cls}-button-checked {
+      color: #fff;
+      background-color: #e6e6e6;
+      border-color: @border-color-base;
+    }
+  }
+}
