+++ components/table/style/index.js
@@ -1,0 +1,8 @@
+import '../../../style/core/index.less';
+import './index.less';
+
+// style dependencies
+import '../../radio/style';
+import '../../checkbox/style';
+import '../../dropdown/style';
+import '../../spin/style';
