+++ components/time-picker/style/index.js
@@ -1,5 +1,2 @@
 import '../../../style/core/index.less';
 import './index.less';
-
-// style dependencies
-import '../../layout/style';
