+++ components/tree-select/style/index.js
@@ -2,4 +2,5 @@ import '../../../style/core/index.less';
 import './index.less';
 
 // style dependencies
-import '../../layout/style';
+import '../../select/style';
+import '../../checkbox/style';
