+++ components/tree/style/index.js
@@ -2,4 +2,4 @@ import '../../../style/core/index.less';
 import './index.less';
 
 // style dependencies
-import '../../layout/style';
+import '../../checkbox/style';
