+++ site/common/lib.js
@@ -1,3 +1,4 @@
+import 'antd/style/index.less';
 import './styles/highlight.less';
 import './styles/common.less';
 import './styles/markdown.less';
