+++ style/components.less
@@ -1,0 +1,41 @@
+@import "../components/affix/style/index";
+@import "../components/alert/style/index";
+@import "../components/badge/style/index";
+@import "../components/breadcrumb/style/index";
+@import "../components/button/style/index";
+@import "../components/calendar/style/index";
+@import "../components/card/style/index";
+@import "../components/carousel/style/index";
+@import "../components/cascader/style/index";
+@import "../components/checkbox/style/index";
+@import "../components/collapse/style/index";
+@import "../components/date-picker/style/index";
+@import "../components/dropdown/style/index";
+@import "../components/form/style/index";
+@import "../components/input/style/index";
+@import "../components/input-number/style/index";
+@import "../components/layout/style/index";
+@import "../components/menu/style/index";
+@import "../components/message/style/index";
+@import "../components/modal/style/index";
+@import "../components/notification/style/index";
+@import "../components/pagination/style/index";
+@import "../components/popover/style/index";
+@import "../components/progress/style/index";
+@import "../components/radio/style/index";
+@import "../components/rate/style/index";
+@import "../components/select/style/index";
+@import "../components/slider/style/index";
+@import "../components/spin/style/index";
+@import "../components/steps/style/index";
+@import "../components/switch/style/index";
+@import "../components/table/style/index";
+@import "../components/tabs/style/index";
+@import "../components/tag/style/index";
+@import "../components/time-picker/style/index";
+@import "../components/timeline/style/index";
+@import "../components/tooltip/style/index";
+@import "../components/transfer/style/index";
+@import "../components/tree/style/index";
+@import "../components/tree-select/style/index";
+@import "../components/upload/style/index";
