+++ style/core/iconfont.less
@@ -1,5 +1,3 @@
-@import "./mixin";
-
 // font-face
 // @icon-url： 字体源文件的地址
 @font-face {
