+++ style/mixins/index.less
@@ -1,16 +1,8 @@
 // Mixins
 // --------------------------------------------------
-
-// Utilities
 @import "opacity.less";
 @import "size.less";
 @import "compatibility";
 @import "clearfix";
 @import "center-block";
-
-// iconfont
 @import "iconfont";
-
-// for common elements
-@import "input.less";
-@import "form.less";
